style(test): preserve event-store evidence anchors
